+++ Assets/ProceduralStages/CHANGELOG.md
@@ -3,6 +3,7 @@
 1. Improved themes
 1. Fixed DCCS pool selecting the same ennemies multiple times
 1. Air node generation is now ~20x faster
+1. Marching cube ~2x faster
 
 ### 1.12